services/customers-ms/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
pika==1.3.2
django-cors-headers==4.3.1
